max_bs_finder/requirements.txt
nvidia-ml-py
onnxruntime-gpu
matplotlib
numpy